Remove leftover chat open call
